// types/dashboard/matrimonial-agreement.ts
export interface MatrimonialAgreementFormValues {
  agreementObjectives: string;
  separatePropertyIntention: string;
  futureAcquisitionsTreatment: string;
  spousalMaintenanceIntention: string;
  childProvisionsIntention: string;
  reviewTriggerEvents: string[];
  reviewIntervalYears: string;
  governingLawConfirmed: boolean;
  separateLegalAdviceUnderstood: boolean;
  specialArrangements: string;
}